+++ foliant/preprocessors/confluence_final/process.py
@@ -80,6 +80,8 @@ def process_code_blocks(content: str, config: dict, chapter=None) -> str:
             language = match.group('syntax')
 
         source = match.group('content')
+        if source.endswith('\n'):
+            source = source[:-1]
         logger.debug(f'Found code block ({language}):\n{source[:150]}\n...')
         return gen_code_macro(source,
                               language=language,
@@ -135,7 +137,7 @@ def gen_code_macro(source: str,
         result += '  <ac:parameter ac:name="linenumbers">true</ac:parameter>\n'
     if collapse:
         result += '  <ac:parameter ac:name="collapse">true</ac:parameter>\n'
-    result += f'<ac:plain-text-body><![CDATA[{source}\n]]></ac:plain-text-body>\n</ac:structured-macro></p></raw_confluence>'
+    result += f'<ac:plain-text-body><![CDATA[{source}]]></ac:plain-text-body>\n</ac:structured-macro></p></raw_confluence>'
     return result